// web/src/api/generated/model/accountInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AccountRole } from './accountRole';
import type { AccountType } from './accountType';

export interface AccountInput {
  code: string;
  nameAr: string;
  nameEn: string;
  /** @nullable */
  parentId: string | null;
  type: AccountType;
  isPosting: boolean;
  role: AccountRole;
}